Update console operator and connectors to use latest pkg and separate gotohelm
